Try overriding sauce credentials after JWT and before sauce connect
