+++ build-system/sauce-credentials.json
@@ -1,10 +1,6 @@
 {
   "_comment": "Encoded per-user Sauce labs credentials. Add yours to this file.",
   "credentials": {
-    "amphtml": {
-      "username": "amphtml",
-      "access_key_encoded": "foobar"
-    },
     "rsimha@ampproject.org": {
       "username": "rsimha",
       "access_key_encoded": "MDg2ZjY4MzQtMWRkZi00YTUyLTgwMWUtMmUwODIyNDY2OTc0Cg=="
